unicorne: Remove Tab mod-tap
Mod-tap keys impose a slight delay on the tapped key. For some keys,
this is fine, but it gets quite annoying for tab which is often used for
autocompletion, where delays get noticed. The only potential downside is
not having a left hand control key, but this is resolved by having it on
the nav layer.